Prevent compiler warnings (well, most of).
Unfortunately, code smell mostly comes from third-party sources,
like Jasper and Unrar, which may be re-imported again in the future.

Warnings that were left:

* Use of `tmpnam` in Jasper.
* Possible wrapping of `localtime_r` for multi-threaded wxWidgets.

+++ plugins/grib_pi/src/jasper/jpc/jpc_mqenc.c
@@ -385,8 +385,8 @@ int jpc_mqenc_dump(jpc_mqenc_t *mqenc, FILE *out)
 {
 	fprintf(out, "AREG = %08x, CREG = %08x, CTREG = %" PRIuFAST32 "\n",
 	  (unsigned int)mqenc->areg, (unsigned int)mqenc->creg, mqenc->ctreg);
-	fprintf(out, "IND = %02d, MPS = %d, QEVAL = %04x\n",
-	  *mqenc->curctx - jpc_mqstates, (*mqenc->curctx)->mps,
+	fprintf(out, "IND = %02u, MPS = %d, QEVAL = %04x\n",
+	  (unsigned int)(*mqenc->curctx - jpc_mqstates), (*mqenc->curctx)->mps,
 	  (unsigned int)(*mqenc->curctx)->qeval);
 	return 0;
 }
